+++ Konsole/ConcurrentWriter.cs
@@ -226,5 +226,19 @@ namespace Konsole
                 _window.MoveBufferArea(sourceLeft,sourceTop,sourceWidth,sourceHeight,targetLeft,targetTop,sourceChar,sourceForeColor,sourceBackColor);
             }
         }
+
+        public ConsoleKeyInfo ReadKey(bool intercept = false)
+        {
+            lock (_locker)
+            {
+                return _window.ReadKey(intercept);
+            }
+            
+        }
+
+        public void KeyWait(params ConsoleKey[] c)
+        {
+            _window.KeyWait(c);
+        }
     }
 }