+++ Konsole/IReadKey.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Konsole
-{
-    public interface IReadKey
-    {
-        ConsoleKeyInfo ReadKey();
-    }
-}